backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.36
asyncpg==0.29.0
alembic==1.13.3
pydantic==2.9.2
python-dotenv==1.0.1
langchain==0.3.7
langchain-groq==0.2.1
langchain-core==0.3.18
langgraph==0.2.50
httpx==0.27.2